adding check availability

+++ server/booking-service/src/main/java/com/evently/booking/ticket/TicketService.java
@@ -1,6 +1,7 @@
 package com.evently.booking.ticket;
 
 import com.evently.booking.exceptions.InsufficientTicketException;
+import com.evently.booking.ticket.entities.TicketMapper;
 import dtos.TicketAllocation;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
@@ -45,26 +46,10 @@ public class TicketService {
     }
 
 
-    public void checkAvailability(int locationEventsId, int ticketCounts) {
-
-
+    public void checkAvailability(long locationEventsId, int ticketCounts) {
         boolean available = ticketRepository.hasAvailableSeats(locationEventsId, ticketCounts);
         if (!available) {
             throw new InsufficientTicketException(locationEventsId, ticketCounts);
         }
-
-        // if available then we will reserve the tickets for the user and then we will ask the user to make the payment and if the payment is successful then we will make the tickets active and if the payment is not successful then we will make the tickets available again.
-        // check avaailnaility
-        // unfortunately, we cannot guarantee the availability of the tickets at this point because of the time gap between checking and booking, so we will have to check the availability again at the time of booking and if the tickets are not available then we will have to inform the user that the tickets are not available and ask them to try again later.
-        //user click book
-        //1. create the tickets
-        // assing the user
-        // assing status
-        // active = false;
-        //2. create the payment
-        //3. if payment successfull then active = true;
-        // ticketStays = SOLD
-        // CRON JOBS CHECKS FOR THE RESERVED TICKETS AND IF THE PAYMENT IS NOT DONE THEN MAKE THE TICKETS AVAILABLE AGAIN
-
     }
 }
